+++ include/nstd/gl.h
@@ -1,6 +1,7 @@
 #ifndef NSTD_GL_H_INCLUDED
 #define NSTD_GL_H_INCLUDED
 #include "gl/buffer.h"
+#include "gl/command.h"
 #include "gl/def.h"
 #include "gl/device.h"
 #include "gl/instance.h"