+++ include/nstd/gl/command.h
@@ -1,0 +1,6 @@
+#ifndef NSTD_GL_COMMAND_H_INCLUDED
+#define NSTD_GL_COMMAND_H_INCLUDED
+#include "../nstd.h"
+#include "command/buffer.h"
+#include "command/encoder.h"
+#endif