+++ include/nstd/gl/command/encoder.h
@@ -1,0 +1,26 @@
+#ifndef NSTD_GL_COMMAND_ENCODER_H_INCLUDED
+#define NSTD_GL_COMMAND_ENCODER_H_INCLUDED
+#include "../../core/def.h"
+#include "../../nstd.h"
+#include "../device/device.h"
+#include "buffer.h"
+NSTDCPPSTART
+
+/// A command encoder for a graphics device.
+typedef NSTDAny NSTDGLCommandEncoder;
+
+/// Creates a gpu command encoder.
+/// Parameters:
+///     `const NSTDGLDevice device` - The device to create commands for.
+/// Returns: `NSTDGLCommandEncoder command_encoder` - The new gpu command encoder.
+NSTDAPI NSTDGLCommandEncoder nstd_gl_command_encoder_new(const NSTDGLDevice device);
+
+/// Frees a command encoder and returns a command buffer.
+/// Parameters:
+///     `NSTDGLCommandEncoder *const command_encoder` - The gpu command encoder.
+/// Returns: `NSTDGLCommandBuffer command_buffer` - The gpu command buffer.
+NSTDAPI NSTDGLCommandBuffer nstd_gl_command_encoder_finish(
+    NSTDGLCommandEncoder *const command_encoder);
+
+NSTDCPPEND
+#endif